requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
playwright
pandas